+++ CHANGELOG.md
@@ -9,7 +9,12 @@
 * which is bad forgetting cloing " ", put evil spaces inside?
 
 
-# 0.3.21 (2025-10-28)
+# 0.3.23-SNAPSHOT
+
+- bug - can not send ratings today is wednesday.
+- fixed - added util/last-tuesday, replaced `today` in switch.clj with it.
+
+# 0.3.22 (2025-10-28)
 
 - added `post /admin/delete` - put! {:db/id eid, :wil2 "delete"}